+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.3 (in development)
 * Chat history now supports images.
+* Images supported by the code interpreter are no longer uploaded twice.
 
 ## 0.1.2 (2025-06-19)
 * Add support for OpenAI's web search capability.